+++ src/main/java/hello/fclover/domain/Goods.java
@@ -6,7 +6,6 @@ import lombok.Getter;
 import lombok.Setter;
 import lombok.ToString;
 
-import java.math.BigInteger;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 
@@ -15,9 +14,9 @@ import java.time.LocalDateTime;
 @ToString
 @Data
 public class Goods {
-    private BigInteger goodsNo;        // 상품 일련번호
-    private BigInteger sellerNo;       // 판매자 번호
-    private Integer cateNo;            // 카테고리 번호
+    private Long goodsNo;        // 상품 일련번호
+    private Long sellerNo;       // 판매자 번호
+    private int cateNo;            // 카테고리 번호
     private String goodsName;          // 상품 명
     private String goodsContent;       // 상품 설명
     private int goodsPrice;            // 상품 가격
